LC-coin-change-ii [Runtime: N/A; Memory: N/A]

+++ complete-knapsack/coin-change-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int change(int amount, vector<int>& coins) {
-        vector<int> dp(amount + 1, 0);
+        vector<long long> dp(amount + 1, 0);
 
         dp[0] = 1;
